// YerbowoFrontend/src/app/account/models/addressCreate.ts
export interface AddressCreate {
    userId: number;
    alias: string;
    firstName: string;
    lastName: string;
    street: string;
    buildingNumber: string;
    apartmentNumber: string;
    place: string;
    postCode: string;
    phone: string;
    email: string;
    nip: string;
    company: string;
}